fixing seeder
