+++ src/routes/index.js
@@ -2,70 +2,12 @@ const { Router } = require("express");
 const axios = require("axios");
 // Importar todos los routers;
 // Ejemplo: const authRouter = require('./auth.js');
-const { Breed, Temperament, BreedTemperament } = require("../db");
+const { Breed, Temperament } = require("../db");
 
 const router = Router();
 
 // Configurar los routers
 // Ejemplo: router.use('/auth', authRouter);
-
-const getApiData = async () => {
-  const apiURL = await axios.get("https://api.thedogapi.com/v1/breeds");
-
-  // console.log(
-  //   apiURL.data[0].temperament
-  //     .toString()
-  //     .split(",")
-  //     .map((t) => t.trim())
-  // );
-
-  // const apiData = await apiURL.data.map((i) => ({
-  //   name: i.name,
-  //   height: i.height.metric,
-  //   weight: i.weight.metric,
-  //   life_span: i.life_span,
-  //   image: i.image.url,
-  //   temperament: i.temperament,
-  // }));
-  // console.log(apiData[0]);
-
-  for (let breed of apiURL.data) {
-    let temps = breed.temperament;
-    if (typeof temps === "string") {
-      temps = temps.split(",").map((t) => t.trim());
-    } else {
-      temps = ["Not available"];
-    }
-
-    let currentBreed = {
-      name: breed.name,
-      height: breed.height.metric,
-      weight: breed.weight.metric,
-      life_span: breed.life_span,
-      image: breed.image.url,
-      createdInDb: false,
-    };
-
-    const savedBreed = await Breed.findOrCreate({
-      where: currentBreed,
-      raw: true,
-    });
-
-    // console.log(savedBreed[0].name);
-
-    for (let temp of temps) {
-      let currentTemp = await Temperament.findOrCreate({
-        where: { name: temp },
-        raw: true,
-      });
-      // console.log(currentTemp[0].id);
-      BreedTemperament.findOrCreate({
-        where: { breedId: savedBreed[0].id, temperamentId: currentTemp[0].id },
-      });
-    }
-  }
-};
-getApiData();
 
 const getDbData = async () => {
   //get data from local db
